requirements.txt
speech_recognition==3.8.1
gTTS==2.2.3
openai==0.11.4
pygame==2.0.2
requests==2.26.0
asciimatics==1.13.0
keyboard==0.13.5
python-dotenv==0.19.1
mysql-connector-python==8.0.27